Added Milestone 6
Partially working prototype, entering either the name of a building or
its code in the "from" and "to" textfields draws a straight path between
the two buildings and fits the map to the path. Also, clicking on any
building on the map will display a popup with the building's name.